requirements.txt
asyncio
discord
matplotlib
numpy
pillow